+++ srcs/main/utils_2.c
@@ -6,16 +6,16 @@
 /*   By: jebouche <jebouche@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/05 19:42:17 by jebouche          #+#    #+#             */
-/*   Updated: 2023/08/30 15:56:51 by jebouche         ###   ########.fr       */
+/*   Updated: 2023/08/31 18:02:04 by jebouche         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3D.h"
 
 /*
- * correct_degrees() corrects the float degrees to be between 0 and 360.
+ * correct_degrees() corrects the double degrees to be between 0 and 360.
  */
-float	correct_degrees(float degrees)
+double	correct_degrees(double degrees)
 {
 	if (degrees < 0)
 		degrees += 360;
@@ -25,11 +25,11 @@ float	correct_degrees(float degrees)
 }
 
 /*
- * deg_to_rad() converts the float degrees to radians.
+ * deg_to_rad() converts the double degrees to radians.
  */
-float	deg_to_rad(float degrees)
+double	deg_to_rad(double degrees)
 {
-	float	radians;
+	double	radians;
 
 	radians = degrees * (M_PI / 180);
 	return (radians);
@@ -39,11 +39,11 @@ float	deg_to_rad(float degrees)
  * get_distance() returns the distance between the t_vecotrs player and 
  * wall_hit using the Pythagorean theorem.
  */
-float	get_distance(t_vector *player, t_vector *wall_hit)
+double	get_distance(t_vector *player, t_vector *wall_hit)
 {
-	float	a;
-	float	b;
-	float	c;
+	double	a;
+	double	b;
+	double	c;
 
 	a = player->x - wall_hit->x;
 	b = player->y - wall_hit->y;
